// src/app/models/tipo-de-consulta-exame.ts
export class TipoDeConsultaExame {
}
export interface TipoDeConsultaExameDTO {
    id: number;
    nome: string;
  }
  
  export interface CreateTipoDeConsultaExameDTO {
    nome: string;
  }
  
  export interface UpdateTipoDeConsultaExameDTO {
    id: number;
    nome?: string;
  }
  
  export interface TipoDeConsultaExame {
    id: number;
    nome: string;

  }